Move SimpleNodeValueClient to org.neo4j.storageengine.api.schema
